requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2.5
image==1.5.27
jedi==0.15.1
joblib==0.13.2
numpy==1.17.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
termcolor==1.1.0
traitlets==4.3.2
wcwidth==0.1.7
